Bug fix: maxSolveLoops(ca,dt) = 0
SO decided to remove the MaxSolveLoops parameter from SPDOPTIONS data. This created an issue with vSPD where the maxSolveLoops is based on this parameter. As a result, the maxSolveLoops in vSPD is set to zero and this causes infinite loop in vSPD.

Solution: Modify vSPD code so that if the dtParameter(ca,dt,'maxSolveLoop') is zero --> using the default value of 5.